+++ docs/v3.1/managers/storage_config.md
@@ -294,4 +294,4 @@ This manager was created in Phase 3d Step 6 to consolidate storage variables:
 
 **Key Finding**: **Infrastructure manager** similar to ServerConfigManager - provides essential storage infrastructure for entire system
 
-**Next Manager**: unified_config_manager.py
+**Next Manager**: unified_config.py
