+++ docs/v3.1/managers/unified_config.md
@@ -1,6 +1,6 @@
 # Unified Config Manager Documentation
 
-**File**: `managers/unified_config_manager.py`  
+**File**: `managers/unified_config.py`  
 **Phase**: 3e Step 1.1 Documentation Audit  
 **Status**: 🔄 **IN PROGRESS**  
 **Factory Function**: `create_unified_config_manager(config_dir)`  
@@ -238,7 +238,7 @@ Infrastructure Layer (File System, Environment)
 
 ## ✅ **Phase 3e Step 1.1 Status**
 
-**Manager**: unified_config_manager.py  
+**Manager**: unified_config.py  
 **Documentation**: ✅ **COMPLETE**  
 **Core Methods**: 25+ critical foundation methods  
 **Shared Methods**: **❌ NONE** - Foundation layer, never extract
