+++ docs/v3.1/phase/3/e/6.1_manager_renaming.md
@@ -16,14 +16,14 @@
 | `managers/analysis_parameters_manager.py` | `managers/analysis_config.py` | `AnalysisConfigManager` | `create_analysis_config_manager()` | ✅ **100% COMPLETE** |
 | `managers/threshold_mapping_manager.py` | `managers/crisis_threshold.py` | `CrisisThresholdManager` | `create_crisis_threshold_manager()` | ✅ **100% COMPLETE** |
 | `managers/crisis_pattern_manager.py` | `managers/pattern_detection.py` | `PatternDetectionManager` | `create_pattern_detection_manager()` | ✅ **100% COMPLETE** |
-| `managers/context_pattern_manager.py` | `managers/context_analysis.py` | `ContextAnalysisManager` | `create_context_analysis_manager()` | 🔄 **NEXT** |
-| `managers/model_ensemble_manager.py` | `managers/model_coordination.py` | `ModelCoordinationManager` | `create_model_coordination_manager()` | ⏳ **PENDING** |
+| `managers/context_pattern_manager.py` | `managers/context_analysis.py` | `ContextAnalysisManager` | `create_context_analysis_manager()` | ✅ **100% COMPLETE** |
+| `managers/model_ensemble_manager.py` | `managers/model_coordination.py` | `ModelCoordinationManager` | `create_model_coordination_manager()` | ✅ **100% COMPLETE** |
 
 ### **Managers Requiring Filename-Only Updates (7 managers):**
 
 | Current File | New File | Class Name | Factory Function | Status |
 |--------------|----------|------------|------------------|--------|
-| `managers/unified_config_manager.py` | `managers/unified_config.py` | `UnifiedConfigManager` | `create_unified_config_manager()` | ⏳ **PENDING** |
+| `managers/unified_config_manager.py` | `managers/unified_config.py` | `UnifiedConfigManager` | `create_unified_config_manager()` | 🔄 **NEXT** |
 | `managers/shared_utilities_manager.py` | `managers/shared_utilities.py` | `SharedUtilitiesManager` | `create_shared_utilities_manager()` | ⏳ **PENDING** |
 | `managers/learning_system_manager.py` | `managers/learning_system.py` | `LearningSystemManager` | `create_learning_system_manager()` | ⏳ **PENDING** |
 | `managers/feature_config_manager.py` | `managers/feature_config.py` | `FeatureConfigManager` | `create_feature_config_manager()` | ⏳ **PENDING** |
