+++ managers/settings.py
