+++ managers/zero_shot.py
